fix(Json Form Link): move to help content
