+++ libs/jsonforms-components/src/lib/Additional/HelpContent.tsx
@@ -6,6 +6,8 @@ import { HelpContentDiv } from './styled-components';
 import { ControlProps, ControlElement } from '@jsonforms/core';
 import { withJsonFormsControlProps } from '@jsonforms/react';
 import { Visible } from '../util';
+import { RenderLink } from './LinkControl';
+
 interface OptionProps {
   ariaLabel?: string;
   help?: string | string[];
@@ -14,6 +16,7 @@ interface OptionProps {
   alt?: string;
   height?: string;
   width?: string;
+  link?: string;
 }
 
 interface CustomControlElement extends ControlElement {
@@ -33,6 +36,7 @@ export const HelpContentComponent = ({
   const marginClass = isParent ? 'parent-margin' : 'child-margin';
   // eslint-disable-next-line
   const { uischema, visible, label } = props;
+  const link = uischema?.options?.link;
   const renderHelp = () =>
     Array.isArray(uischema?.options?.help) ? (
       <ul>
@@ -59,7 +63,8 @@ export const HelpContentComponent = ({
             </div>
           )}
           {uischema.options?.img && uischema.options?.img !== '' && renderImage(uischema.options)}
-          {(!uischema.options?.variant || uischema.options?.variant !== 'details') && renderHelp()}
+          {uischema?.options?.link && link !== '' && RenderLink(uischema?.options)}
+          {(!uischema.options?.variant || uischema.options?.variant !== 'details') && !link && renderHelp()}
           {uischema.options?.variant && uischema.options?.variant === 'details' && (
             <GoADetails heading={label ? label : ''} mt="3xs" mb="none">
               {renderHelp()}
